pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "discordia-support-italia"
version = "0.1.0"
[tool.poetry.dependencies]
"discord.py" = "^1.3"
flask = "^1.1"
python = "^3.8"
dblpy = "^0.3.4"
[tool.poetry.dev-dependencies]
